fixed main page and other fixes
