+++ server/controllers/jobs.py
@@ -73,6 +73,7 @@ def add_job_to_queue():
         # favorite = 1 if _favorite == 'true' else 0
         quantity = request.form['quantity']
         td_id = int(request.form['td_id'])
+        filament = request.form['filament']
         favoriteOne = False 
 
         for i in range(int(quantity)):
@@ -83,7 +84,7 @@ def add_job_to_queue():
                 favorite = 0
                 
             status = 'inqueue' # set status 
-            res = Job.jobHistoryInsert(name, printer_id, status, file, file_name_original, favorite, td_id) # insert into DB 
+            res = Job.jobHistoryInsert(name, printer_id, status, file, file_name_original, favorite, td_id, filament) # insert into DB 
             
             # retrieve job from DB
             id = res['id']
@@ -120,6 +121,7 @@ def auto_queue():
 
         favorite = request.form['favorite']
         td_id = request.form['td_id']
+        filament = request.form['filament']
 
         favoriteOne = False 
         for i in range(int(quantity)):
@@ -133,7 +135,7 @@ def auto_queue():
                 favorite = 0
             # favorite = 1 if _favorite == 'true' else 0
             
-            res = Job.jobHistoryInsert(name, printer_id, status, file, file_name_original, favorite, td_id) # insert into DB 
+            res = Job.jobHistoryInsert(name, printer_id, status, file, file_name_original, favorite, td_id, filament) # insert into DB 
             
             id = res['id']
             
@@ -568,7 +570,7 @@ def repair_ports():
 def getFilament(): 
     try:
         job_id = request.args.get('jobid', default=-1, type=int)
-        res = Job.getFilament(job_id)
+        res = Job.getFilamentFromJob(job_id)
         return jsonify(res)
     except Exception as e:
         print(f"Unexpected error: {e}")
@@ -590,8 +592,9 @@ def rerunjob(printerpk, jobpk, position):
     file_name_original = job.getFileNameOriginal() # get original file name
     favorite = job.getFileFavorite() # get favorite status
     td_id = job.getTdId() 
+    filament = job.getFilament()
     # Insert new job into DB and return new PK 
-    res = Job.jobHistoryInsert(name=job.getName(), printer_id=printerpk, status=status, file=job.getFile(), file_name_original=file_name_original, favorite = favorite, td_id=td_id) # insert into DB 
+    res = Job.jobHistoryInsert(name=job.getName(), printer_id=printerpk, status=status, file=job.getFile(), file_name_original=file_name_original, favorite=favorite, td_id=td_id, filament=filament) # insert into DB 
     
     id = res['id']
     file_name_pk = file_name_original + f"_{id}" # append id to file name to make it unique
