+++ server/models/PrinterStatusService.py
@@ -156,6 +156,7 @@ class PrinterStatusService:
                     "comments": job.comments, 
                     "extruded": job.extruded,
                     "td_id": job.td_id,
+                    "filament": job.filament,
                 }
                 printer_info['queue'].append(job_info)
